Fix consecutive block ends crashing interp bug
Two codeblocks ending consecutively with no expression in between 
eg
```
if(foo){
    if(bar){
        trace('hello');
    }
    //no expression here would trigger the bug
}
```
Would crash the script. The fetch phase will now, on encountering the end of a block, continue bubbling up through stack frames until it either finds one that will continue or returns a value and ends the script.